backend/requirements.txt
flask==3.1.3
tensorflow==2.21.0
numpy==2.4.4
pillow==12.1.1
h5py==3.14.0
<<<<<<< HEAD
scipy
=======
>>>>>>> e31c6cf8
gunicorn
